Functioned for Stef run

+++ utilities.py
@@ -37,15 +37,15 @@ def directory_contents(directory_path):
 
 # Delete a list of file 
 def delete_files(file_paths):
-     
-    for file_path in file_paths:
-        try:
-            os.remove(f"./image_temp/{file_path}")
-            print(f"Remove : ./image_temp/{file_path}")
-        except OSError as e:
-            print(f"Error: {file_path} : {e.strerror}")
-    file_paths.clear()
-    
+    if file_paths: 
+        for file_path in file_paths:
+            try:
+                os.remove(f"./image_temp/{file_path}")
+                print(f"Remove : ./image_temp/{file_path}")
+            except OSError as e:
+                print(f"Error: {file_path} : {e.strerror}")
+        file_paths.clear()
+        
 
 def open_image_from_url(image_url):
     response = requests.get(image_url)
